Stop leaving defaults in the database (hopefully one last time...)
This commit adds testing to verify that no defaults are left in the
database even through alter-column, and even when the column is not
null, and fixes for all backends  to comply (except SQLite and Oracle,
which complied already).

The fix for Firebird is untested. MySQL also untested, but shares the
code path with Postgres, so hopefully ok.